+++ syndicate/core/generators/project.py
@@ -16,78 +16,91 @@
 
 import os
 
+import yaml
+
 from syndicate.commons.log_helper import get_logger
+from syndicate.core import ProjectState
 from syndicate.core.generators import (_touch, _mkdir,
                                        _write_content_to_file)
 from syndicate.core.generators.contents import (_get_lambda_default_policy,
                                                 JAVA_ROOT_POM_TEMPLATE,
-                                                SRC_MAIN_JAVA, FILE_POM)
-from syndicate.core.groups import (PROJECT_JAVA, PROJECT_NODEJS,
-                                   PROJECT_PYTHON)
+                                                SRC_MAIN_JAVA, FILE_POM,
+                                                CHANGELOG_TEMPLATE,
+                                                GITIGNORE_CONTENT,
+                                                README_TEMPLATE)
+from syndicate.core.groups import (RUNTIME_JAVA, RUNTIME_NODEJS,
+                                   RUNTIME_PYTHON)
 
 _LOG = get_logger('syndicate.core.generators.project')
 
 SLASH_SYMBOL = '/'
-FOLDER_LAMBDAS = '/lambdas'
-FOLDER_COMMONS = '/commons'
-FILE_README = '/README.md'
-FILE_DEPLOYMENT_RESOURCES = '/deployment_resources.json'
+FOLDER_LAMBDAS = 'lambdas'
+FOLDER_COMMONS = 'commons'
+FILE_README = 'README.md'
+FILE_DEPLOYMENT_RESOURCES = 'deployment_resources.json'
+FILE_CHANGELOG = 'CHANGELOG.md'
+FILE_GITIGNORE = '.gitignore'
 
 
-
-def generate_project_structure(project_name, project_path, project_language):
+def generate_project_structure(project_name, project_path):
     try:
         if not os.path.exists(project_path):
             raise AssertionError(
                 'Path "{}" you have provided does not exist'.format(
                     project_path))
 
-        processor = PROJECT_PROCESSORS.get(project_language)
-        if not processor:
-            raise RuntimeError('Wrong project language {0}'.format(
-                project_language))
-
-        full_project_path = project_path + SLASH_SYMBOL + project_name if (
-                project_path[
-                    -1] != SLASH_SYMBOL) else project_path + project_name
+        full_project_path = os.path.join(project_path, project_name) \
+            if (project_path[-1] != SLASH_SYMBOL) \
+            else project_path + project_name
 
         _mkdir(path=full_project_path,
                fault_message='Folder {} already exists. \nOverride the '
                              'project? [y/n]: '.format(full_project_path))
 
-        _touch(full_project_path + FILE_README)
-        default_lambda_policy = _get_lambda_default_policy()
-        _write_content_to_file(full_project_path + FILE_DEPLOYMENT_RESOURCES,
-                               default_lambda_policy)
+        path_to_readme = os.path.join(full_project_path, FILE_README)
+        _touch(path_to_readme)
+        readme_template = README_TEMPLATE.replace('project_name', project_name)
+        _write_content_to_file(file=path_to_readme,
+                               content=readme_template)
 
-        processor(project_name=project_name,
-                  full_project_path=full_project_path)
-        _LOG.info('Project {} has been successfully created.'.format(
+        default_lambda_policy = _get_lambda_default_policy()
+        _write_content_to_file(file=os.path.join(full_project_path,
+                                                 FILE_DEPLOYMENT_RESOURCES),
+                               content=default_lambda_policy)
+        _mkdir(path=os.path.join(full_project_path, 'src'), exist_ok=True)
+        ProjectState.generate(project_name=project_name,
+                              project_path=full_project_path)
+
+        _write_content_to_file(os.path.join(full_project_path, FILE_CHANGELOG),
+                               CHANGELOG_TEMPLATE)
+        _write_content_to_file(os.path.join(full_project_path, FILE_GITIGNORE),
+                               GITIGNORE_CONTENT)
+        _LOG.info('Project {} folder has been successfully created.'.format(
             project_name))
     except Exception as e:
-        _LOG.error(str(e))
+        _LOG.exception(str(e))
 
 
 def _generate_python_project_hierarchy(full_project_path, project_name=None):
-    _mkdir(full_project_path + FOLDER_LAMBDAS, exist_ok=True)
+    _mkdir(os.path.join(full_project_path, FOLDER_LAMBDAS), exist_ok=True)
 
 
 def _generate_java_project_hierarchy(project_name, full_project_path):
-    _touch(full_project_path + FILE_POM)
+    pom_path = os.path.join(full_project_path, FILE_POM)
+    _touch(path=pom_path)
     pom_content = JAVA_ROOT_POM_TEMPLATE.replace('{project_name}',
                                                  project_name)
-    _write_content_to_file(full_project_path + FILE_POM,
-                           pom_content)
+    _write_content_to_file(file=pom_path, content=pom_content)
     _mkdir(full_project_path + SRC_MAIN_JAVA)
 
 
 def _generate_nodejs_project_hierarchy(full_project_path, project_name=None):
-    _mkdir(full_project_path + FOLDER_LAMBDAS, exist_ok=True)
-    _mkdir(full_project_path + FOLDER_COMMONS, exist_ok=True)
+    _mkdir(os.path.join(full_project_path, FOLDER_LAMBDAS), exist_ok=True)
+    _mkdir(os.path.join(full_project_path, FOLDER_COMMONS), exist_ok=True)
 
 
 PROJECT_PROCESSORS = {
-    PROJECT_JAVA: _generate_java_project_hierarchy,
-    PROJECT_NODEJS: _generate_nodejs_project_hierarchy,
-    PROJECT_PYTHON: _generate_python_project_hierarchy,
+    RUNTIME_JAVA: _generate_java_project_hierarchy,
+    RUNTIME_NODEJS: _generate_nodejs_project_hierarchy,
+    RUNTIME_PYTHON: _generate_python_project_hierarchy,
 }