+++ syndicate/core/handlers.py
@@ -18,8 +18,10 @@ import os
 import sys
 
 import click
+from tabulate import tabulate
 
-from syndicate.core import CONF_PATH, initialize_connection
+from syndicate.core import CONF_PATH, initialize_connection, \
+    initialize_project_state
 from syndicate.core.build.artifact_processor import (RUNTIME_NODEJS,
                                                      assemble_artifacts,
                                                      RUNTIME_JAVA_8,
@@ -33,7 +35,13 @@ from syndicate.core.build.deployment_processor import (
     remove_deployment_resources, remove_failed_deploy_resources,
     update_deployment_resources)
 from syndicate.core.build.meta_processor import create_meta
-from syndicate.core.conf.generator import generate_configuration_files
+from syndicate.core.build.profiler_processor import (get_metric_statistics,
+                                                     process_metrics)
+from syndicate.core.build.warmup_processor import (process_deploy_resources,
+                                                   process_api_gw_resources,
+                                                   warm_upper,
+                                                   process_existing_api_gw_id,
+                                                   process_inputted_api_gw_id)
 from syndicate.core.conf.validator import (MVN_BUILD_TOOL_NAME,
                                            PYTHON_BUILD_TOOL_NAME,
                                            NODE_BUILD_TOOL_NAME)
@@ -44,7 +52,14 @@ from syndicate.core.helper import (check_required_param,
                                    handle_futures_progress_bar,
                                    resolve_path_callback, timeit,
                                    verify_bundle_callback,
-                                   verify_meta_bundle_callback)
+                                   verify_meta_bundle_callback,
+                                   resolve_default_value,
+                                   generate_default_bundle_name,
+                                   sync_lock)
+from syndicate.core.project_state.project_state import (MODIFICATION_LOCK,
+                                                        WARMUP_LOCK)
+from syndicate.core.project_state.status_processor import project_state_status
+from syndicate.core.project_state.sync_processor import sync_project_state
 
 INIT_COMMAND_NAME = 'init'
 commands_without_config = (
@@ -61,170 +76,117 @@ def _not_require_config(all_params):
 @click.version_option()
 def syndicate():
     if CONF_PATH:
-        click.echo('Path to sdct.conf: ' + CONF_PATH)
+        click.echo('Configuration used: ' + CONF_PATH)
         initialize_connection()
+        initialize_project_state()
     elif _not_require_config(sys.argv):
         pass
     else:
         click.echo('Environment variable SDCT_CONF is not set! '
                    'Please verify that you configured have provided path to '
-                   'correct config files or execute `syndicate init` command.')
+                   'correct config files '
+                   'or execute `syndicate generate config` command.')
         sys.exit(1)
 
 
-@syndicate.command(name=INIT_COMMAND_NAME)
-@click.option('--config_path', type=str,
-              help='Path to store generated configuration file')
-@click.option('--project_path', type=str,
-              help='Path to project folder. Default value: working dir')
-# account settings
-@click.option('--account_id', callback=check_required_param, type=str,
-              help='[required] Id of the AWS account where to deploy '
-                   'application')
-@click.option('--region', type=str, default='us-west-1',
-              help='The region that is used to deploy the application')
-@click.password_option(
-    '--access_key', type=str,
-    help='AWS access key id that is used to deploy the application.')  # todo get from .aws/credentials/[default] if not specified
-@click.password_option(
-    '--secret_key', type=str,
-    help='AWS secret key that is used to deploy the application.')  # todo get from .aws/credentials/[default] if not specified
-@click.option('--bundle_bucket_name', type=str,
-              help='Name of the bucket that is used for uploading artifacts. '
-                   'It will be created if specified.')
-# build_projects_mapping
-@click.option('--python_build_mapping', '-pbm', multiple=True,
-              help='List of the folders in a project where '
-                   'Python code must be assembled')
-@click.option('--java_build_mapping', '-jbm', multiple=True,
-              help='List of the folders in a project where '
-                   'Java code must be assembled')
-@click.option('--nodejs_build_mapping', '-nbm', multiple=True,
-              help='List of the folders in a project where '
-                   'NodeJS code must be assembled')
-# deploy setting
-@click.option('--prefix', type=str,
-              help='Prefix that is added to project names while deployment '
-                   'by pattern: {prefix}resource_name{suffix}')
-@click.option('--suffix', type=str,
-              help='Suffix that is added to project names while deployment '
-                   'by pattern: {prefix}resource_name{suffix}')
-def init(config_path, project_path, account_id, region, access_key,
-         secret_key, bundle_bucket_name, python_build_mapping,
-         java_build_mapping, nodejs_build_mapping, prefix, suffix):
-    """
-    Configures aws-syndicate:
-    - generates sdct.conf using the provided parameters;
-    - creates default sdct_aliases.conf;
-    :param config_path: path where the generated configurations files
-        will be stored
-    :param project_path: path to a project which is supposed to be deployed
-        by aws-syndicate
-    :param account_id: id of the AWS account where an application
-        will be deployed
-    :param region: AWS region name where an application will be deployed
-    :param access_key: AWS access key id to access an AWS account specified
-        in account_id parameter. If access_key is not specified
-        the credentials from /$user/.aws/credentials will be used if any.
-        The provided credentials should provide enough permissions to deploy
-        the app.
-    :param secret_key: AWS secret access key to access an AWS account
-        specified in account_id parameter. If secret_key is not specified
-        the credentials from /$user/.aws/credentials will be used if any.
-        The provided credentials should provide enough permissions to deploy
-        the app.
-    :param bundle_bucket_name: name of the bucket in specified AWS account
-        where application bundles will be stored
-    :param python_build_mapping: path[s] to the python parts of an application.
-        Example: --python_build_mapping /src/$app_python_module/
-    :param java_build_mapping: path[s] to the java parts of an application.
-        Example: --java_build_mapping /src/$app_java_module/
-    :param nodejs_build_mapping: path[s] to the nodejs parts of an application.
-        Example: --nodejs_build_mapping /src/$node_python_module/
-    :param prefix: will be added to resource names of types
-    iam_role, iam_policy, s3_bucket.
-    Resource name pattern: {prefix}resource_name{suffix}
-    :param suffix: will be added to resource names of types
-    iam_role, iam_policy, s3_bucket.
-    Resource name pattern: {prefix}resource_name{suffix}
-    :return:
-    """
-    generate_configuration_files(config_path=config_path,
-                                 project_path=project_path,
-                                 region=region,
-                                 account_id=account_id,
-                                 access_key=access_key,
-                                 secret_key=secret_key,
-                                 bundle_bucket_name=bundle_bucket_name,
-                                 python_build_mapping=python_build_mapping,
-                                 java_build_mapping=java_build_mapping,
-                                 nodejs_build_mapping=nodejs_build_mapping,
-                                 prefix=prefix,
-                                 suffix=suffix)
+@syndicate.command(name='test')
+@click.option('--suite', type=click.Choice(['unittest', 'pytest', 'nose'],
+                                           case_sensitive=False),
+              default='unittest')
+@click.option('--test_folder_name', nargs=1, default='tests')
+@timeit()
+def test(suite, test_folder_name):
+    """Discovers and runs tests inside python project configuration path."""
+    click.echo('Running tests...')
+    import subprocess
+    from syndicate.core import CONFIG
+    project_path = CONFIG.project_path
+
+    test_folder = os.path.join(project_path, test_folder_name)
+    if not os.path.exists(test_folder):
+        click.echo(f'Tests not found, \'{test_folder_name}\' folder is missing'
+                   f' in \'{project_path}\'.')
+        return
+
+    test_lib_command_mapping = {
+        'unittest': 'python -m unittest -v',
+        'pytest': 'pytest --no-header -v',
+        'nose': 'nosetests --verbose'
+    }
+
+    workdir = os.getcwd()
+
+    os.chdir(test_folder)
+    command = test_lib_command_mapping.get(suite)
+    result = subprocess.run(command.split())
+
+    os.chdir(workdir)
+    if result.returncode != 0:
+        click.echo('Some tests failed. Exiting.')
+        sys.exit(1)
 
 
-@syndicate.command(name='build_bundle')
-@click.option('--bundle_name', nargs=1, callback=check_required_param)
-@click.option('--force_upload', is_flag=True, default=False)
+@syndicate.command(name='build')
+@click.option('--bundle_name', nargs=1,
+              callback=generate_default_bundle_name,
+              help='Name of the bundle to build. '
+                   'Default value: $ProjectName_%Y-%m-%dT%H:%M:%SZ')
+@click.option('--force_upload', is_flag=True, default=False,
+              help='Flag to override existing bundle with the same name')
 @click.pass_context
-@timeit
-def build_bundle(ctx, bundle_name, force_upload):
+@timeit(action_name='build')
+def build(ctx, bundle_name, force_upload):
     """
     Builds bundle of an application
-    :param ctx:
-    :param bundle_name: name of the bundle
-    :param force_upload: used to override existing bundle
-    :return:
     """
     if if_bundle_exist(bundle_name=bundle_name) and not force_upload:
         click.echo('Bundle name \'{0}\' already exists '
                    'in deploy bucket. Please use another bundle '
                    'name or delete the bundle'.format(bundle_name))
         return
-    ctx.invoke(build_artifacts, bundle_name=bundle_name)
+    ctx.invoke(test)
+    ctx.invoke(assemble, bundle_name=bundle_name)
     ctx.invoke(package_meta, bundle_name=bundle_name)
-    ctx.invoke(upload_bundle, bundle_name=bundle_name, force=force_upload)
+    ctx.invoke(upload, bundle_name=bundle_name, force=force_upload)
 
 
 @syndicate.command(name='deploy')
-@click.option('--deploy_name', nargs=1, callback=check_required_param)
-@click.option('--bundle_name', nargs=1, callback=check_required_param)
-@click.option('--deploy_only_types', multiple=True)
-@click.option('--deploy_only_resources', multiple=True)
-@click.option('--deploy_only_resources_path', nargs=1)
-@click.option('--excluded_resources', multiple=True)
-@click.option('--excluded_resources_path', nargs=1)
-@click.option('--excluded_types', multiple=True)
-@click.option('--continue_deploy', is_flag=True)
-@click.option('--replace_output', nargs=1, is_flag=True, default=False)
+@click.option('--deploy_name',
+              callback=resolve_default_value,
+              help='Name of the deploy. Default value: name of the project')
+@click.option('--bundle_name',
+              callback=resolve_default_value,
+              help='Name of the bundle to deploy. '
+                   'Default value: name of the latest built bundle')
+@click.option('--deploy_only_types', multiple=True,
+              help='Types of the resources to deploy')
+@click.option('--deploy_only_resources', multiple=True,
+              help='Names of the resources to deploy')
+@click.option('--deploy_only_resources_path', nargs=1,
+              help='Path to file containing names of the resources to deploy')
+@click.option('--excluded_resources', multiple=True,
+              help='Names of the resources to skip while deploy.')
+@click.option('--excluded_resources_path', nargs=1,
+              help='Path to file containing names of the resources to skip '
+                   'while deploy')
+@click.option('--excluded_types', multiple=True,
+              help='Types of the resources to skip while deploy')
+@click.option('--continue_deploy', is_flag=True,
+              help='Flag to continue failed deploy')
+@click.option('--replace_output', is_flag=True, default=False,
+              help='Replaces the existing deploy output')
 @check_deploy_name_for_duplicates
-@timeit
+@sync_lock(lock_type=MODIFICATION_LOCK)
+@timeit(action_name='deploy')
 def deploy(deploy_name, bundle_name, deploy_only_types, deploy_only_resources,
            deploy_only_resources_path, excluded_resources,
            excluded_resources_path, excluded_types, continue_deploy,
            replace_output):
     """
-    Deploys infrastructure from the specified bundle
-    :param deploy_name: name of the deploy
-    :param bundle_name: name of the bundle
-    :param deploy_only_types: list of types of the resources to deploy
-    :param deploy_only_resources: list of resources names to deploy
-    :param deploy_only_resources_path: path to a json file that contains
-        a list of resources names to deploy
-    :param excluded_resources: names of the resources which must be
-        skipped while deploy
-    :param excluded_resources_path: path to a json file that contains a list
-        of resources names which must be skipped while deploy
-    :param excluded_types: list of types of resources which must be
-        skipped while deploy
-    :param continue_deploy: continues deploy using the failed output.
-        Used only after previous deploy fail.
-    :param replace_output: flag to override the output file.
-        Used if previous output file must be overridden.
-    :return:
+    Deploys the application infrastructure
     """
-    click.echo('Command deploy backend')
-    click.echo('Deploy name: %s' % deploy_name)
+    sync_project_state()
+    from syndicate.core import PROJECT_STATE
     if deploy_only_resources_path and os.path.exists(
             deploy_only_resources_path):
         deploy_resources_list = json.load(open(deploy_only_resources_path))
@@ -252,31 +214,34 @@ def deploy(deploy_name, bundle_name, deploy_only_types, deploy_only_resources,
                                                      replace_output)
     click.echo('Backend resources were deployed{0}.'.format(
         '' if deploy_success else ' with errors. See deploy output file'))
+    PROJECT_STATE.release_lock(MODIFICATION_LOCK)
+    sync_project_state()
 
 
 @syndicate.command(name='update')
-@click.option('--bundle_name', nargs=1, callback=check_required_param)
-@click.option('--deploy_name', nargs=1, callback=check_required_param)
-@click.option('--update_only_types', multiple=True)
-@click.option('--update_only_resources', multiple=True)
-@click.option('--update_only_resources_path', nargs=1)
+@click.option('--bundle_name',
+              callback=resolve_default_value,
+              help='Name of the bundle to deploy. '
+                   'Default value: name of the latest built bundle')
+@click.option('--deploy_name',
+              callback=resolve_default_value,
+              help='Name of the deploy. Default value: name of the project')
+@click.option('--update_only_types', multiple=True,
+              help='Types of the resources to update')
+@click.option('--update_only_resources', multiple=True,
+              help='Names of the resources to deploy')
+@click.option('--update_only_resources_path', nargs=1,
+              help='Path to file containing names of the resources to skip '
+                   'while deploy')
 @click.option('--replace_output', nargs=1, is_flag=True, default=False)
 @check_deploy_name_for_duplicates
-@timeit
+@timeit(action_name='update')
 def update(bundle_name, deploy_name, replace_output,
            update_only_resources,
            update_only_resources_path,
            update_only_types=[]):
     """
     Updates infrastructure from the provided bundle
-    :param bundle_name: name of the bundle
-    :param deploy_name: name of the deploy
-    :param update_only_resources: list of resources names to updated
-    :param update_only_resources_path: path to a json file with list of
-        resources names to update
-    :param update_only_types: optional. List of a resources types to update.
-    :param replace_output: flag. If True, existing output file will be replaced
-    :return:
     """
     click.echo('Bundle name: {}'.format(bundle_name))
     if update_only_types:
@@ -304,9 +269,9 @@ def update(bundle_name, deploy_name, replace_output,
 
 
 @syndicate.command(name='clean')
-@timeit
-@click.option('--deploy_name', nargs=1, callback=check_required_param)
-@click.option('--bundle_name', nargs=1, callback=check_required_param)
+@timeit(action_name='clean')
+@click.option('--deploy_name', nargs=1, callback=resolve_default_value)
+@click.option('--bundle_name', nargs=1, callback=resolve_default_value)
 @click.option('--clean_only_types', multiple=True)
 @click.option('--clean_only_resources', multiple=True)
 @click.option('--clean_only_resources_path', nargs=1, type=str)
@@ -320,42 +285,24 @@ def clean(deploy_name, bundle_name, clean_only_types, clean_only_resources,
           excluded_resources_path, excluded_types, rollback):
     """
     Cleans the application infrastructure.
-    :param deploy_name: name of the deploy
-    :param bundle_name: name of the bundle
-    :param clean_only_types: list of types that must be cleaned
-    :param clean_only_resources: list of names of resources that
-        must be cleaned
-    :param clean_only_resources_path: path to a json list which contains a list
-        of resources names which must be cleaned
-    :param clean_externals: used to clean external resources
-    :param excluded_resources: list of the resources names than must be skipped
-        while cleaning the application infrastructure
-    :param excluded_resources_path: path to a json list which contains a list
-        of resources names which must be skipped while cleaning
-        the application infrastructure
-    :param excluded_types: list of types that must be skipped while cleaning
-        the application infrastructure
-    :param rollback: used to clean resources that were created in scope of the
-        failed deploy
-    :return:
     """
     click.echo('Command clean')
-    click.echo('Deploy name: %s' % deploy_name)
+    click.echo(f'Deploy name: {deploy_name}')
+    separator = ', '
     if clean_only_types:
-        click.echo('Clean only types: %s' % str(clean_only_types))
+        click.echo(f'Clean only types: {separator.join(clean_only_types)}')
     if clean_only_resources:
-        click.echo('Clean only resources : %s' % clean_only_resources)
+        click.echo(f'Clean only resources: '
+                   f'{separator.join(clean_only_resources)}')
     if clean_only_resources_path:
-        click.echo(
-            'Clean only resources path: %s' % clean_only_resources_path)
+        click.echo(f'Clean only resources path: {clean_only_resources_path}')
     if excluded_resources:
-        click.echo('Excluded resources: %s' % str(excluded_resources))
+        click.echo(f'Excluded resources: {separator.join(excluded_resources)}')
     if excluded_resources_path:
-        click.echo('Excluded resources path: %s' % excluded_resources_path)
+        click.echo(f'Excluded resources path: {excluded_resources_path}')
     if excluded_types:
-        click.echo('Excluded types: %s' % str(excluded_types))
-    if clean_only_resources_path and os.path.exists(
-            clean_only_resources_path):
+        click.echo(f'Excluded types: {separator.join(excluded_resources)}')
+    if clean_only_resources_path and os.path.exists(clean_only_resources_path):
         clean_resources_list = json.load(open(clean_only_resources_path))
         clean_only_resources = tuple(
             set(clean_only_resources + tuple(clean_resources_list)))
@@ -382,11 +329,110 @@ def clean(deploy_name, bundle_name, clean_only_types, clean_only_resources,
     click.echo('AWS resources were removed.')
 
 
+@syndicate.command(name='sync')
+@timeit()
+def sync():
+    """
+    Syncs the state of local project state file (.syndicate) and
+    the remote one.
+    """
+    return sync_project_state()
+
+
+@syndicate.command(name='status')
+@click.option('--events', 'category', flag_value='events',
+              help='Show event logs of the project')
+@click.option('--resources', 'category', flag_value='resources',
+              help='Show a summary of the project resources')
+@timeit()
+def status(category):
+    """
+    Shows the state of a local project state file (.syndicate).
+    """
+    click.echo(project_state_status(category))
+
+
+@syndicate.command(name='warmup')
+@click.option('--bundle_name', nargs=1, callback=resolve_default_value,
+              help='Name of the bundle. Should be specified with deploy_name'
+                   ' parameter.')
+@click.option('--deploy_name', nargs=1, callback=resolve_default_value,
+              help='Name of the deploy.')
+@click.option('--api_gw_id', nargs=1, multiple=True, type=str,
+              help='Provide API Gateway IDs to warmup.')
+@click.option('--stage_name', nargs=1, multiple=True, type=str,
+              help='Name of stages of provided API Gateway IDs.')
+@click.option('--lambda_auth', default=False, is_flag=True,
+              help='Should be specified if API Gateway Lambda Authorizer is '
+                   'enabled')
+@click.option('--header_name', nargs=1, help='Name of authentication header.')
+@click.option('--header_value', nargs=1, help='Name of authentication header '
+                                              'value.')
+@sync_lock(lock_type=WARMUP_LOCK)
+@timeit(action_name='warmup')
+def warmup(bundle_name, deploy_name, api_gw_id, stage_name, lambda_auth,
+           header_name, header_value):
+    """
+    Warmups Lambda functions.
+    """
+
+    if bundle_name and deploy_name:
+        click.echo(f'Deploy name: {deploy_name}')
+        if not if_bundle_exist(bundle_name=bundle_name):
+            click.echo(f'Bundle name \'{bundle_name}\' does not exists '
+                       'in deploy bucket. Please use another bundle '
+                       'name or create the bundle')
+            return
+
+        paths_to_be_triggered, resource_path_warmup_key_mapping = \
+            process_deploy_resources(deploy_name=deploy_name,
+                                     bundle_name=bundle_name)
+
+    elif api_gw_id:
+        paths_to_be_triggered, resource_path_warmup_key_mapping = \
+            process_inputted_api_gw_id(api_id=api_gw_id, stage_name=stage_name,
+                                       echo=click.echo)
+
+    else:
+        paths_to_be_triggered, resource_path_warmup_key_mapping = \
+            process_existing_api_gw_id(stage_name=stage_name, echo=click.echo)
+
+    resource_method_mapping, resource_warmup_key_mapping = \
+        process_api_gw_resources(paths_to_be_triggered=paths_to_be_triggered,
+                                 resource_path_warmup_key_mapping=
+                                 resource_path_warmup_key_mapping)
+    warm_upper(resource_method_mapping=resource_method_mapping,
+               resource_warmup_key_mapping=resource_warmup_key_mapping,
+               lambda_auth=lambda_auth, header_name=header_name,
+               header_value=header_value)
+    click.echo('Application resources have been warmed up.')
+
+
+@syndicate.command(name='profiler')
+@click.option('--bundle_name', nargs=1, callback=resolve_default_value)
+@click.option('--deploy_name', nargs=1, callback=resolve_default_value)
+@click.option('--from_date', nargs=1, type=str)
+@click.option('--to_date', nargs=1, type=str)
+def profiler(bundle_name, deploy_name, from_date, to_date):
+    """
+    Displays application Lambda metrics
+    """
+    metric_value_dict = get_metric_statistics(bundle_name, deploy_name,
+                                              from_date, to_date)
+    for lambda_name, metrics in metric_value_dict.items():
+        prettify_metrics_dict = {}
+
+        click.echo(f'{os.linesep}Lambda function name: {lambda_name}')
+        prettify_metrics_dict = process_metrics(prettify_metrics_dict, metrics)
+        click.echo(tabulate(prettify_metrics_dict, headers='keys',
+                            stralign='right'))
+
+
 # =============================================================================
 
 
 @syndicate.command(name='assemble_java_mvn')
-@timeit
+@timeit()
 @click.option('--bundle_name', nargs=1, callback=create_bundle_callback)
 @click.option('--project_path', '-path', nargs=1,
               callback=resolve_path_callback)
@@ -405,7 +451,7 @@ def assemble_java_mvn(bundle_name, project_path):
 
 
 @syndicate.command(name='assemble_python')
-@timeit
+@timeit()
 @click.option('--bundle_name', nargs=1, callback=create_bundle_callback)
 @click.option('--project_path', '-path', nargs=1,
               callback=resolve_path_callback)
@@ -424,7 +470,7 @@ def assemble_python(bundle_name, project_path):
 
 
 @syndicate.command(name='assemble_node')
-@timeit
+@timeit()
 @click.option('--bundle_name', nargs=1, callback=create_bundle_callback)
 @click.option('--project_path', '-path', nargs=1,
               callback=resolve_path_callback)
@@ -449,11 +495,11 @@ COMMAND_TO_BUILD_MAPPING = {
 }
 
 
-@syndicate.command(name='build_artifacts')
-@timeit
+@syndicate.command(name='assemble')
+@timeit()
 @click.option('--bundle_name', nargs=1, callback=create_bundle_callback)
 @click.pass_context
-def build_artifacts(ctx, bundle_name):
+def assemble(ctx, bundle_name):
     """
     Builds the application artifacts
     :param ctx:
@@ -462,22 +508,22 @@ def build_artifacts(ctx, bundle_name):
     :return:
     """
     click.echo('Building artifacts ...')
-    from syndicate.core import CONFIG
-    if CONFIG.build_projects_mapping:
-        for key, values in CONFIG.build_projects_mapping.items():
-            for value in values:
-                func = COMMAND_TO_BUILD_MAPPING.get(key)
-                if func:
-                    ctx.invoke(func, bundle_name=bundle_name,
-                               project_path=value)
-                else:
-                    click.echo('Build tool is not supported: %s' % key)
+    from syndicate.core import PROJECT_STATE
+    build_mapping_dict = PROJECT_STATE.load_project_build_mapping()
+    if build_mapping_dict:
+        for key, value in build_mapping_dict.items():
+            func = COMMAND_TO_BUILD_MAPPING.get(key)
+            if func:
+                ctx.invoke(func, bundle_name=bundle_name,
+                           project_path=value)
+            else:
+                click.echo('Build tool is not supported: %s' % key)
     else:
         click.echo('Projects to be built are not found')
 
 
 @syndicate.command(name='package_meta')
-@timeit
+@timeit()
 @click.option('--bundle_name', nargs=1, callback=verify_bundle_callback)
 def package_meta(bundle_name):
     """
@@ -493,7 +539,7 @@ def package_meta(bundle_name):
 
 
 @syndicate.command(name='create_deploy_target_bucket')
-@timeit
+@timeit()
 def create_deploy_target_bucket():
     """
     Creates a bucket in AWS account where all bundles will be uploaded
@@ -505,11 +551,11 @@ def create_deploy_target_bucket():
     click.echo('Deploy target bucket was created successfully')
 
 
-@syndicate.command(name='upload_bundle')
-@timeit
+@syndicate.command(name='upload')
 @click.option('--bundle_name', nargs=1, callback=verify_meta_bundle_callback)
 @click.option('--force', is_flag=True)
-def upload_bundle(bundle_name, force=False):
+@timeit(action_name='upload')
+def upload(bundle_name, force=False):
     """
     Uploads bundle from local storage to AWS S3
     :param bundle_name: name of the bundle to upload
@@ -520,8 +566,10 @@ def upload_bundle(bundle_name, force=False):
     click.echo('Upload bundle: %s' % bundle_name)
     if force:
         click.echo('Force upload')
+
     futures = upload_bundle_to_s3(bundle_name=bundle_name, force=force)
     handle_futures_progress_bar(futures)
+
     click.echo('Bundle was uploaded successfully')
 
 
@@ -536,7 +584,7 @@ def upload_bundle(bundle_name, force=False):
 @click.option('--role_name', '-role', nargs=1,
               callback=check_required_param)
 @click.option('--force_upload', is_flag=True, default=False)
-@timeit
+@timeit()
 @click.pass_context
 def copy_bundle(ctx, bundle_name, src_account_id, src_bucket_region,
                 src_bucket_name, role_name, force_upload):
@@ -563,7 +611,7 @@ def copy_bundle(ctx, bundle_name, src_account_id, src_bucket_region,
                           src_bucket_name, role_name)
     handle_futures_progress_bar(futures)
     click.echo('Bundle was downloaded successfully')
-    ctx.invoke(upload_bundle, bundle_name=bundle_name, force=force_upload)
+    ctx.invoke(upload, bundle_name=bundle_name, force=force_upload)
     click.echo('Bundle was copied successfully')